# Fix `gpt_bpe` concurrency performance, implement `StatusWatcher`