--- a/cmd/dataset_tokenizer/dataset_tokenizer.go
+++ b/cmd/dataset_tokenizer/dataset_tokenizer.go
@@ -33,7 +33,7 @@ type PathInfo struct {
 // returning a slice of PathInfo.
 func GlobTexts(dirPath string) (pathInfos []PathInfo, err error) {
 	// If the path is a file, return it.
-	if stat, statErr := os.Stat(dirPath); statErr != nil && !stat.IsDir() {
+	if stat, _ := os.Stat(dirPath); stat != nil && !stat.IsDir() {
 		return []PathInfo{{
 			Path:    dirPath,
 			Size:    stat.Size(),