--- a/cmd/dataset_tokenizer/go.mod
+++ b/cmd/dataset_tokenizer/go.mod
@@ -1,8 +1,6 @@
 module github.com/wbrown/gpt_bpe/cmd/dataset_tokenizer
 
-go 1.22
-
-toolchain go1.22.6
+go 1.19
 
 //replace github.com/wbrown/gpt_bpe => ../../