--- a/cmd/dataset_tokenizer/go.mod
+++ b/cmd/dataset_tokenizer/go.mod
@@ -2,8 +2,6 @@ module github.com/wbrown/gpt_bpe/cmd/dataset_tokenizer
 
 go 1.19
 
-//replace github.com/wbrown/gpt_bpe => ../../
-
 require (
 	github.com/aws/aws-sdk-go v1.45.4
 	github.com/stretchr/testify v1.9.0