Following changes were made:
* Add recognition and acceptance of `.md` files.
* Move reporting into a `StatusWatcher` gothread that periodically reports rather than spam the console with output per-artifact.

```
2024/10/20 17:21:38 46006600 tokens in 10.00s, 0 pad tokens, 4600638.26 tokens/s, 332 docs, avg 138574.10 tokens/doc
2024/10/20 17:21:38   Thread 0: [sent, waiting], 11624537 tokens, 0 pad tokens, 90 docs, Times: 9.858s tokenizing, 2ms recv, 0s send, File: James Clavell - Shogun.txt
2024/10/20 17:21:38   Thread 1: [sent, waiting], 11460692 tokens, 0 pad tokens, 85 docs, Times: 10.062s tokenizing, 1ms recv, 0s send, File: Colleen McCullough [Masters of Rome 4] Caesars Women.txt
2024/10/20 17:21:38   Thread 2: [tokenizing: inchsz=16, outchsz=2, idx: 21910/46590], 11542605 tokens, 0 pad tokens, 78 docs, Times: 10.082s tokenizing, 0s recv, 0s send, File: Robert A. Heinlein - Orphans of the Sky.txt
2024/10/20 17:21:38   Thread 3: [tokenizing: inchsz=16, outchsz=1, waiting], 11378766 tokens, 0 pad tokens, 79 docs, Times: 10.037s tokenizing, 0s recv, 0s send, File: William Gibson - Neuromancer.txt
```

* Implement `GPT_Encoder.Clone()` which makes a shallow copy, with the most important separate state, the LRU cache, being reinstantiated.
* Each thread gets its own `GPT_Encoder` instance so that the LRU cache mutex doesn't block the multiple threads.
* The above gets us past 2m tokens a second ceiling that we were hitting due to the LRU cache being shared.
* Move some instantiation logic into a separate function that is only called once. This takes care of race conditions.
* Some formatting changes and cleanup.